pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "biucingcli"
version = "0.1.0"
description = "Full-stack developer toolchain command-line assistant."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "BiucingCLI Maintainers" }
]
keywords = ["cli", "toolchain", "developer-tools", "typer", "automation"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
]
dependencies = [
  "typer[all]>=0.12.3",
  "rich>=13.7.0",
  "pyyaml>=6.0.1",
]

[project.urls]
Homepage = "https://example.com/biucingcli"
Repository = "https://example.com/biucingcli/repo"

[project.optional-dependencies]
dev = [
  "black>=24.4.2",
  "ruff>=0.6.4",
  "mypy>=1.11.1",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
]

[project.scripts]
biucing = "biucingcli.cli:app"

[tool.hatch.version]
path = "src/biucingcli/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
  "src/biucingcli",
  "README.md",
  "CHANGELOG.md",
  "LICENSE",
  "pyproject.toml",
]

[tool.hatch.build.targets.wheel]
packages = ["src/biucingcli"]

[tool.hatch.envs.default]
dependencies = [
  "black>=24.4.2",
  "ruff>=0.6.4",
  "mypy>=1.11.1",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
]

[tool.hatch.envs.default.scripts]
format = "black src tests"
lint = "ruff check src tests"
typer-check = "mypy src"
test = "pytest --cov=biucingcli --cov-report=term-missing"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "PL", "RUF"]
ignore = ["E203", "E501"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-ra"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict = true
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["biucingcli"]

[tool.coverage.report]
show_missing = true
skip_covered = true
